requirements.txt
torchvision  # 請用 pip show torchvision 確認你的版本
torchaudio   # 請用 pip show torchaudio 確認你的版本
accelerate
tensorboard
pymeshlab
huggingface_hub
diffusers
xformers
jaxtyping
omegaconf
typeguard
einops
opencv-python
imageio
matplotlib
kiui
icecream
fpsample
plyfile
open3d
numpy
